Extract duplicated logic to MapGeneratorBase trait.
