+++ OpenRA.Mods.Common/Traits/World/ClearMapGenerator.cs
@@ -9,9 +9,6 @@
  */
 #endregion
 
-using System.Collections.Generic;
-using System.Collections.Immutable;
-using System.Linq;
 using OpenRA.Mods.Common.MapGenerator;
 using OpenRA.Support;
 using OpenRA.Traits;
@@ -20,100 +17,17 @@ namespace OpenRA.Mods.Common.Traits
 {
 	[TraitLocation(SystemActors.EditorWorld)]
 	[Desc("A map generator that clears a map.")]
-	public sealed class ClearMapGeneratorInfo : TraitInfo, IEditorMapGeneratorInfo
+	public sealed class ClearMapGeneratorInfo : MapGeneratorBaseInfo
 	{
-		[FieldLoader.Require]
-		[Desc("Human-readable name this generator uses.")]
-		[FluentReference]
-		public readonly string Name = null;
+		protected override int GetPlayerCount(MapGenerationArgs args) => 0;
 
-		[FieldLoader.Require]
-		[Desc("Internal id for this map generator.")]
-		public readonly string Type = null;
-
-		[FieldLoader.Require]
-		[Desc("Tilesets that are compatible with this map generator.")]
-		public readonly ImmutableArray<string> Tilesets = default;
-
-		[FluentReference]
-		[Desc("The title to use for generated maps.")]
-		public readonly string MapTitle = "label-random-map";
-
-		[Desc("The widget tree to open when the tool is selected.")]
-		public readonly string PanelWidget = "MAP_GENERATOR_TOOL_PANEL";
-
-		// This is purely of interest to the linter.
-		[FieldLoader.LoadUsing(nameof(LoadFluentReferences))]
-		[FluentReference]
-		public readonly ImmutableArray<string> FluentReferences = default;
-
-		[FieldLoader.LoadUsing(nameof(LoadOptions))]
-		public readonly ImmutableArray<MapGeneratorOption> Options;
-
-		string IMapGeneratorInfo.Type => Type;
-		string IMapGeneratorInfo.Name => Name;
-		string IMapGeneratorInfo.MapTitle => MapTitle;
-
-		static object LoadOptions(MiniYaml my)
-		{
-			var optionsNode = my.NodeWithKeyOrDefault("Options");
-			if (optionsNode == null)
-				return ImmutableArray<MapGeneratorOption>.Empty;
-
-			var options = new List<MapGeneratorOption>();
-			foreach (var node in optionsNode.Value.Nodes)
-			{
-				var split = node.Key.Split('@');
-				if (split.Length != 2)
-					continue;
-
-				if (split[0] == "BooleanOption")
-					options.Add(new MapGeneratorBooleanOption(split[1], node.Value));
-				else if (split[0] == "IntegerOption")
-					options.Add(new MapGeneratorIntegerOption(split[1], node.Value));
-				else if (split[0] == "MultiIntegerChoiceOption")
-					options.Add(new MapGeneratorMultiIntegerChoiceOption(split[1], node.Value));
-				else if (split[0] == "MultiChoiceOption")
-					options.Add(new MapGeneratorMultiChoiceOption(split[1], node.Value));
-			}
-
-			return options.ToImmutableArray();
-		}
-
-		static object LoadFluentReferences(MiniYaml my)
-		{
-			return ((ImmutableArray<MapGeneratorOption>)LoadOptions(my))
-				.SelectMany(o => o.GetFluentReferences())
-				.ToImmutableArray();
-		}
-
-		MiniYaml GenerateParameterYaml(ModData modData, MapGenerationArgs args)
-		{
-			var terrainInfo = modData.DefaultTerrainInfo[args.Tileset];
-
-			// Apply the choices in their canonical order.
-			var parameters = new Dictionary<string, MiniYaml>();
-			foreach (var o in Options.OrderBy(option => option.Priority))
-			{
-				if (!args.Options.TryGetValue(o.Id, out var value))
-					continue;
-
-				foreach (var pn in o.GetParameters(terrainInfo, value, 0))
-					parameters[pn.Key] = pn.Value;
-			}
-
-			return new MiniYaml(null, parameters.Select(kv => new MiniYamlNode(kv.Key, kv.Value)));
-		}
-
-		public Map Generate(ModData modData, MapGenerationArgs args)
+		public override Map Generate(ModData modData, MapGenerationArgs args)
 		{
 			var random = new MersenneTwister();
-			var terrainInfo = modData.DefaultTerrainInfo[args.Tileset];
-			var yaml = GenerateParameterYaml(modData, args);
-
-			if (!Exts.TryParseUshortInvariant(yaml.NodeWithKey("Tile").Value.Value, out var tileType))
+			if (!Exts.TryParseUshortInvariant(GenerateParameterYaml(modData, args).NodeWithKey("Tile").Value.Value, out var tileType))
 				throw new YamlException("Illegal tile type");
 
+			var terrainInfo = modData.DefaultTerrainInfo[args.Tileset];
 			if (!terrainInfo.TryGetTerrainInfo(new TerrainTile(tileType, 0), out _))
 				throw new MapGenerationException("Illegal tile type");
 
@@ -130,36 +44,15 @@ namespace OpenRA.Mods.Common.Traits
 			return map;
 		}
 
-		public bool TryGenerateMetadata(ModData modData, MapGenerationArgs args, out MapPlayers players, out Dictionary<string, MiniYaml> ruleDefinitions)
-		{
-			ruleDefinitions = [];
-			players = new MapPlayers(modData.DefaultRules, 0);
-
-			return true;
-		}
-
 		public override object Create(ActorInitializer init)
 		{
-			return new ClearMapGenerator(this);
+			return new ClearMapGenerator(init, this);
 		}
-
-		ImmutableArray<string> IEditorMapGeneratorInfo.Tilesets => Tilesets;
-		ImmutableArray<MapGeneratorOption> IEditorMapGeneratorInfo.Options => Options;
-		int IEditorMapGeneratorInfo.GetPlayerCount(MapGenerationArgs args) => 0;
 	}
 
-	public class ClearMapGenerator : IEditorTool
+	public class ClearMapGenerator : MapGeneratorBase
 	{
-		public string Label { get; }
-		public string PanelWidget { get; }
-		public TraitInfo TraitInfo { get; }
-		public bool IsEnabled => true;
-
-		public ClearMapGenerator(ClearMapGeneratorInfo info)
-		{
-			Label = info.Name;
-			PanelWidget = info.PanelWidget;
-			TraitInfo = info;
-		}
+		public ClearMapGenerator(ActorInitializer init, ClearMapGeneratorInfo info)
+			: base(init, info) { }
 	}
 }
